Refine async lifecycle ownership and organization
