+++ evalscope/perf/core/__init__.py
@@ -1,5 +1,6 @@
 from evalscope.perf.core.http_client import AioHttpClient, test_connection
-from evalscope.perf.core.metrics_consumer import connect_test, run_benchmark_pipeline, statistic_benchmark_metric
+from evalscope.perf.core.metrics_consumer import connect_test, statistic_benchmark_metric
+from evalscope.perf.core.pipeline import run_benchmark_pipeline
 from evalscope.perf.core.strategies import BenchmarkStrategy, ClosedLoopStrategy, MultiTurnStrategy, OpenLoopStrategy
 
 __all__ = [
